Springboot Rest API - Security Day 1

+++ src/main/java/com/juaracoding/repo/GroupMenuRepo.java
@@ -3,5 +3,5 @@ package com.juaracoding.repo;
 import com.juaracoding.model.GroupMenu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupMenuRepo extends JpaRepository<GroupMenu, Integer> {
+public interface GroupMenuRepo extends JpaRepository<GroupMenu, Long> {
 }
